backend/requirements.txt
Django
django-environ
djoser
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
Pillow
channels
channels-redis
selenium